Sigfault Solved

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: aalfahal <aalfahal@student.42abudhabi.ae>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/08 15:55:43 by aalfahal          #+#    #+#             */
-/*   Updated: 2023/04/02 05:56:22 by aalfahal         ###   ########.fr       */
+/*   Updated: 2023/04/04 07:41:21 by aalfahal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ size_t	ft_strlen(const char *s)
 	return (i);
 }
 
-size_t	ft_strlen_2d(char **s)
+int	ft_strlen_2d(char **s)
 {
 	int	i;
